refactor: 컴포넌트를 app/components로 이동
