+++ packages/web/src/app/components/ExtensionDialog/index.tsx
@@ -1,5 +1,9 @@
 'use client';
 
+import { EXTENSION, URL_CHROME_STORE, URL_GUIDE_KO } from '@extension/shared/constants';
+import useTranslation from '@src/app/i18n/client';
+import { LanguageType } from '@src/app/i18n/type';
+import { Button } from '@src/components/ui/button';
 import {
   Dialog,
   DialogContent,
@@ -8,75 +12,91 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@src/components/ui/dialog';
-
-import { useEffect, useState } from 'react';
 import { useGetExtensionManifest } from '@src/hooks';
-import { URL_CHROME_STORE } from '@extension/shared/constants';
-import { Button } from '@src/components/ui/button';
-import { LocalStorage, LOCAL_STORAGE_KEY_MAP } from '@src/utils';
-
-const EXTENSION_DIALOG = {
-  install: {
-    message: {
-      title: '확장 프로그램 미설치',
-      description: '설치해야 메모 기능을 이용하실 수 있습니다. 설치하러 가시겠습니까?',
-      cancel: '설치하지 않고 이용하기',
-      ok: '설치하러 가기',
-    },
-    link: URL_CHROME_STORE,
-    localStorage: LOCAL_STORAGE_KEY_MAP.install,
-  },
-  update: {
-    message: {
-      title: '새로운 버전 출시!',
-      description: '',
-      cancel: '업데이트 하지 않고 이용하기',
-      ok: '업데이트 하러가기',
-    },
-    link: '',
-    localStorage: LOCAL_STORAGE_KEY_MAP.updateVersion,
-  },
-};
+import { checkUpdateVersion, LOCAL_STORAGE_KEY_MAP, LocalStorage } from '@src/utils';
+import { useEffect, useState } from 'react';
 
 type DialogType = 'install' | 'update';
 
-const VERSION = '1.6.3';
+interface ExtensionDialogProps extends LanguageType {}
 
-export default function ExtensionDialog() {
+export default function ExtensionDialog({ lng }: ExtensionDialogProps) {
+  const { t } = useTranslation(lng);
   const [open, setOpen] = useState(false);
   const manifest = useGetExtensionManifest();
+
   const [dialogType, setDialogType] = useState<DialogType | null>(null);
 
-  console.log(manifest);
   useEffect(() => {
     if (manifest === null) return;
-    if (manifest === undefined) setDialogType('install');
-    else if (manifest.version !== VERSION) setDialogType('update');
+
+    const isExtensionInstalled = manifest !== undefined;
+    const isExtensionNotLastVersion = isExtensionInstalled && manifest.version !== EXTENSION.lastVersion;
+
+    if (!isExtensionInstalled && !LocalStorage.check(LOCAL_STORAGE_KEY_MAP.install)) {
+      setDialogType('install');
+      setOpen(true);
+    } else if (isExtensionNotLastVersion && !LocalStorage.check(LOCAL_STORAGE_KEY_MAP.updateVersion)) {
+      setDialogType('update');
+      setOpen(true);
+    }
   }, [manifest]);
 
   if (!dialogType) return;
 
-  const handleUpdateClick = () => {
-    window.open(EXTENSION_DIALOG[dialogType].link, '_blank');
+  const EXTENSION_DIALOG = {
+    install: {
+      message: {
+        title: t('dialogInstall.title'),
+        description: t('dialogInstall.title'),
+        ok: t('dialogInstall.ok'),
+        cancel: t('dialogInstall.cancel'),
+      },
+      link: URL_CHROME_STORE,
+      localStorageKey: LOCAL_STORAGE_KEY_MAP.install,
+    },
+    update: {
+      message: {
+        title: t('dialogVersion.title'),
+        description: t('dialogVersion.description', {
+          currentVersion: manifest?.version,
+          lastVersion: EXTENSION.lastVersion,
+        }),
+        cancel: t('dialogVersion.cancel'),
+        ok: t('dialogVersion.ok'),
+      },
+      link: URL_GUIDE_KO,
+      localStorageKey: LOCAL_STORAGE_KEY_MAP.updateVersion + 'a',
+    },
   };
 
-  const handleCloseClick = () => {
+  const currentExtensionDialog = EXTENSION_DIALOG[dialogType];
+
+  const setLocalStroageTrueAndCloseDialog = () => {
     setOpen(false);
-    LocalStorage.setTrue(EXTENSION_DIALOG[dialogType].localStorage);
+    if (checkUpdateVersion(currentExtensionDialog.localStorageKey))
+      LocalStorage.setTrue(currentExtensionDialog.localStorageKey);
+  };
+
+  const handleUpdateClick = () => {
+    window.open(currentExtensionDialog.link, '_blank');
+    setLocalStroageTrueAndCloseDialog();
   };
 
   return (
-    <Dialog open={!!dialogType} onOpenChange={setOpen}>
+    <Dialog open={open} onOpenChange={setOpen}>
       <DialogContent>
         <DialogHeader>
-          <DialogTitle>{EXTENSION_DIALOG[dialogType].message.title}</DialogTitle>
-          <DialogDescription>{EXTENSION_DIALOG[dialogType].message.description}</DialogDescription>
+          <DialogTitle>{currentExtensionDialog.message.title}</DialogTitle>
+          <DialogDescription>{currentExtensionDialog.message.description}</DialogDescription>
         </DialogHeader>
         <DialogFooter>
-          <Button onClick={handleCloseClick} variant="secondary">
-            {EXTENSION_DIALOG[dialogType].message.cancel}
-          </Button>
-          <Button onClick={handleUpdateClick}>{EXTENSION_DIALOG[dialogType].message.ok}</Button>
+          {currentExtensionDialog.message.cancel && (
+            <Button onClick={setLocalStroageTrueAndCloseDialog} variant="secondary">
+              {currentExtensionDialog.message?.cancel}
+            </Button>
+          )}
+          <Button onClick={handleUpdateClick}>{currentExtensionDialog.message.ok}</Button>
         </DialogFooter>
       </DialogContent>
     </Dialog>
