+++ packages/web/src/app/components/ExtensionDialog/useExtensionDialog.ts
@@ -1,0 +1,7 @@
+import { useState } from 'react';
+
+export default function useExtensionDialog() {
+  const [state, setState] = useState();
+
+  return {};
+}
