+++ packages/web/src/components/index.ts
@@ -5,4 +5,3 @@ export { default as QueryProvider } from './QueryProvider';
 export { default as HydrationBoundaryWrapper } from './HydrationBoundaryWrapper';
 export { default as ThemeProvider } from './ThemeProvider';
 export { default as ToggleTheme } from './ToggleTheme';
-export { default as ExtensionDialog } from './ExtensionDialog';
